# MentorLink/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MentorLink API",
        "description": "API RESTful MentorLink — Auth via Bearer token (Sanctum)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/sessions/{sessionId}/reviews": {
            "post": {
                "tags": [
                    "Évaluations"
                ],
                "summary": "Déposer une évaluation",
                "operationId": "fbd4bdaadfaaebcc82db2b0026571218",
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Évaluation enregistrée"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "409": {
                        "description": "Évaluation déjà déposée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mentors/{mentorId}/reviews": {
            "get": {
                "tags": [
                    "Évaluations"
                ],
                "summary": "Évaluations d'un mentor",
                "operationId": "b3aae669bab4c16bfcad29bedfe48a3c",
                "parameters": [
                    {
                        "name": "mentorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des reviews"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}